API-650: Move client in VO
